integrations/api/requirements_external.txt
# CWB Hub External API - Requirements
# Task 16 - API para Integração com Sistemas Externos

# FastAPI e dependências
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Autenticação e Segurança
PyJWT>=2.8.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
cryptography>=41.0.7

# Rate Limiting e Cache
redis>=5.0.1
hiredis>=2.2.3

# HTTP Client
httpx>=0.25.2
aiohttp>=3.9.0

# Utilitários
python-dotenv>=1.0.0
python-dateutil>=2.8.2
typing-extensions>=4.8.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Logging e Monitoramento
structlog>=23.2.0
prometheus-client>=0.19.0
sentry-sdk[fastapi]>=1.38.0

# Database (opcional)
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# File Processing
aiofiles>=23.2.1
python-magic>=0.4.27

# Validation e Serialization
marshmallow>=3.20.0
cerberus>=1.3.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=20.1.0

# Development Tools
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.0
mkdocstrings[python]>=0.24.0

# Performance
aiocache>=0.12.2
orjson>=3.9.0

# Utilities
click>=8.1.7
rich>=13.6.0
tqdm>=4.66.0

# Email (opcional)
aiosmtplib>=3.0.0
email-validator>=2.1.0